refactor global memory
