+++ crates/cubecl-matmul/src/components/global/global_memory/mod.rs
@@ -1,0 +1,7 @@
+mod reader;
+mod tma;
+mod writer;
+
+pub use reader::{TensorReader, Window};
+pub use tma::MappedTensorReader;
+pub use writer::TensorWriter;
