+++ crates/cubecl-matmul/src/components/global/load/loader/async_buffer_loader.rs
@@ -2,7 +2,7 @@ use super::StageIdent;
 use crate::components::global::base::GlobalConfig;
 use crate::components::global::load::{AsyncLoadingJob, LoadingValidation};
 use crate::components::global::multi_stage::double_buffering::DoubleBufferingGlobalConfig;
-use crate::components::global::tensor_view::TensorReader;
+use crate::components::global::global_memory::TensorReader;
 use crate::components::global::{CopyMechanism, Quantization};
 use crate::components::stage::PartialStageToTileReader;
 use crate::components::stage::TilingLayout;
