+++ crates/cubecl-matmul/src/components/global/load/loader/async_full_loader.rs
@@ -2,7 +2,7 @@ use std::marker::PhantomData;
 
 use crate::components::global::Quantization;
 use crate::components::global::load::{AsyncLoadingJob, LoadingValidation};
-use crate::components::global::tensor_view::TensorReader;
+use crate::components::global::global_memory::TensorReader;
 use crate::components::global::{CopyMechanism, GlobalConfig};
 use crate::components::stage::FullStageToTileReader;
 use crate::components::stage::TilingLayout;
