+++ crates/cubecl-matmul/src/components/global/load/loader/sync_buffer_loader.rs
@@ -11,7 +11,7 @@ use crate::components::global::load::LoadingValidation;
 use crate::components::global::multi_stage::Job;
 use crate::components::global::multi_stage::JobExecutor;
 use crate::components::global::multi_stage::LoadMaxRoundPlaneCount;
-use crate::components::global::tensor_view::TensorReader;
+use crate::components::global::global_memory::TensorReader;
 use crate::components::stage::PartialStageToTileReader;
 use crate::components::stage::StageMemory;
 use crate::components::stage::TilingLayout;
