+++ crates/cubecl-matmul/src/components/global/load/strategy/async_full_cooperative.rs
@@ -3,7 +3,7 @@ use crate::components::{
     global::{
         CopyMechanism, GlobalConfig,
         load::AsyncFullLoadingStrategy,
-        tensor_view::{TensorReader, Window},
+        global_memory::{TensorReader, Window},
     },
     stage::{StageMemory, StridedTilingLayout},
 };
