+++ crates/cubecl-matmul/src/components/global/load/strategy/async_full_maximize_unit_count.rs
@@ -3,7 +3,7 @@ use crate::components::{
     global::{
         CopyMechanism, GlobalConfig,
         load::AsyncFullLoadingStrategy,
-        tensor_view::{TensorReader, Window},
+        global_memory::{TensorReader, Window},
     },
     stage::{StageConfig, StageMemory, StridedTilingLayout},
 };
