+++ crates/cubecl-matmul/src/components/global/load/strategy/base.rs
@@ -1,4 +1,4 @@
-use crate::components::global::tensor_view::TensorReader;
+use crate::components::global::global_memory::TensorReader;
 use crate::components::global::{CopyMechanism, GlobalConfig, Quantization};
 use crate::components::stage::{StageMemory, TilingLayout};
 use crate::components::{Ident, InvalidConfigError, MatmulPrecision};
