+++ crates/cubecl-matmul/src/components/global/load/strategy/sync_buffer_cyclic.rs
@@ -2,7 +2,7 @@ use std::marker::PhantomData;
 
 use crate::components::global::load::SyncBufferLoadingStrategy;
 use crate::components::global::multi_stage::LoadMaxRoundPlaneCount;
-use crate::components::global::tensor_view::TensorReader;
+use crate::components::global::global_memory::TensorReader;
 use crate::components::global::{GlobalConfig, Quantization, RoleRule};
 use crate::components::stage::{ContiguousTilingLayout, StageMemory, TilingOrder};
 use crate::components::{Ident, InputIdent, InvalidConfigError, MatmulPrecision, TilingScheme};
