+++ crates/cubecl-matmul/src/components/global/write/plane.rs
@@ -1,6 +1,6 @@
 use crate::components::Ident;
 use crate::components::global::GlobalConfig;
-use crate::components::global::tensor_view::TensorWriter;
+use crate::components::global::global_memory::TensorWriter;
 use cubecl_core as cubecl;
 use cubecl_core::prelude::*;
 use cubecl_std::div_ceil;
